LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-redirect-push-expected.txt


FAIL precommitHandler redirect() push assert_equals: expected "#redirect2" but got "#push"
